requirements.txt
Authlib==1.5.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
pycparser==2.22
python-dotenv==1.0.1
python-json-logger==3.3.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
